Change JavacBasicAnnotationProcessor to wrap each XProcessingStep individually.
This commit changes the behavior of JavacBasicAnnotationProcessor to wrap each XProcessingStep individually in its own Step rather than have a single Step that wraps all XProcessingSteps. The issue with wrapping all XProcessingSteps in a single Step is that the underlying BasicAnnotationProcessor triggers the Step, and thus all wrapped XProcessingSteps, when any annotation is found rather than only triggering the Step that handles the annotation.

In order to share the XProcessingEnv for all XProcessingSteps in a round, I've added a Step at the beginning of each round that initializes the XProcessingEnv for all subsequent rounds. In addition, this also allows the postRound call to share the same XProcessingEnv as the previous Steps in that round. At the end of each postRound I've set the XProcessingEnv to null to allow GC.