Fix video stream freeze.
- MutableLiveData.postValue jumps to the main thread before reporting
the state change. This means the activity does not receive both connecting and
connected states if the bound happens before the thread jump.
For example, in case of resume the activity never receives the connecting state and
only gets the connected state eventhough view model tries to report the
connecting state.

It is hard to verify if this change addresses the root cause of the
issue however, here is the hypothetis behind this fin. We could not
repro the issue after the fix.

- When activity resumes it recreates a surface and reports the surface
to the host since the listener is never cleared out.
The activity then finishes the initialization and receives a listener from
the host and reports its surface to the host for the second time.
The host replaces the old presentation with the new one. However,
when the old presentation gets collected it releases the surface
connection resulting in a broken connection even for the new
presentation since they both share the same surface.

Test: Manual
Change-Id: I887f2db757053ba8f715e9bf4f5e033135ea36e3