Support top level declarations in libraries
This CL adds support for top level members in compiled code.

When a top level function/property is compiled, Kotlin generates a
synthetic class to contain it. Unfortunately, that class is not
accessible in the KSP type system as a ClassDeclaration.

XProcessing already supported files as member containers and this CL
adds support for files that are already compiled (hence no KSFile).

We synthesize a member container for them
(KspSyntheticFileMemberContainer).

I've also added a constant for buganizer link and cleaned up all
references in xprocessing (not in room compiler though).

Fixes: 188822146
Test: TopLevelMembersTest
Change-Id: Id5bced8c35a9bec74bcf8a15e5c365505aa47fa9