Desktop. Make DesktopPlatform internal, introduce internal MouseScrollableConfig
Fixes https://github.com/JetBrains/compose-jb/issues/303

The original intention of this issue was to introduce DesktopLookAndFeel
as changing DesktopPlatform in runtime can bring confusion.

But it appears that that changing scrollable behaviour by only predefined enum's
isn't flexible, so this CL introduces `interface MouseScrollableConfig`
(internal for now, as we don't know the shape of the new Pointer Input API yet).

DesktopPlatform also made internal, because it is probably not a good API:
- the choices are limited
- there can be problems in the future if we add a new platform
- platforms are not only differentiated by OS, but also by architecture/shell/etc

Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true
Change-Id: I207fcd4f2f3faf162fd960bfabf29ab5dff9aad1