Move complication highlight state into RenderParameters
We also fix handling of ambient in CanvasComplicationDrawable which
needs to be set based on RenderParameter's drawMode.

Test: Manual testing and new golden image test.
Bug: 193111096
Relnote: CanvasComplication.isHighlighted has been moved into RenderParameters.pressedComplicationSlotIds this is a step towards making CanvasComplication stateless. To support this change CanvasComplication.render now also takes the slotId as a parameter and we now pass the ComplicationSlot into GlesTextureComplication.
Change-Id: I50e6e600d10f512689b680404eda737595b7a5e8