Implementing a check for verifying that equals() and hashCode() are implemented by the key of a multimap query result.
The function implementsEqualsAndHashcode() takes in the XType representing the type argument of the map key, and recursively looks for the equals() and hashCode() methods among the declared methods of type element class. If they are not found at the current class level, the method recursively moves on to the super class level and continues to look for these declared methods until they are either found, or the method reaches the topmost level (the Object class) in which case the method returns false.

Bug: 187490856
Test: QueryMethodProcessorTest.kt
Relnote: Added a check for verifying that equals() and hashCode() are implemented by the key of a multimap query result.
Change-Id: I828b0612735f0d2702226b4e6238e857b3fbd7f0