First versions of Wear Material List and Lazy list to get consensus on names and API
Tests for these components will come in follow up patches/patchsets once the API shape and naming are resolved.
Kudos to andreykulikov@ for the LazyList implementation.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 193388276
Bug: 193796372
Bug: 193796160

RelNote: "Adding ScalingLazyColumn component that provides a list components for Wear Material that give a fisheye view with the list contents scaling down in size and becoming transparent as they scale towards the edge of the component."

Change-Id: I7070cd6a7250352dc4d9c0a110a9e7468dc02c0a