Add a minimal java sample that has complications.
This commit adds a sample that contains a single complications and has
a matching configuration activity. This is meant to be a minimal
sample that only showcases this aspect of the API. It is also written
in Java, so that it makes use of the Guava wrappers for the various
Kotlin APIs, which we expect a number of our developers to use.

Test: manually ran the sample
Change-Id: Ic64f29f3a700fa74eb09d23e729d0bfab7e7ae3e