Always call updateSelection if selection has changed
Previously we only call InputMethodManager.updateSelection when the
EditBuffer is out of sync with current TextFieldValue. And when the
selection is updated through text edit, we didn't call
updateSelection.
From the doc of the InputMethodManager
//https://developer.android.com/reference/android/view/inputmethod/InputMethodManager#updateSelection(android.view.View,%20int,%20int,%20int,%20int)
and the implementation of EditText in Android framework updateSelection
should be called even though the selection update is from IMM.

Bug: 193457105
Test: ./gradlew compose:ui:ui:connectedAndroidTest
Change-Id: Ie11891f0de9449bcf08f60c9232a02758e3135da