Update back handling when popping from a Navigator
When a Navigator (such as a DialogNavigator)
initiates a pop, we update back stack without
going through `popBackStack()`, which means our
popBackStackFromNavigator needs to handle all
of the work that is normally done for us.

This includes updating the enabled state of
the OnBackPressedCallback - this is particularly
important when the dialog is stacked directly
on top of the start destination - popping the
dialog should set the callback to disabled.

Relnote: "Fixed an issue where popping a
dialog destination would not update the NavController's
system back button handling, potentially causing
the NavController to intercept the back button
even though it does not have any back stack to pop."
Test: tested in Navigation Compose Basic Nav Demo

Change-Id: If3b3464fea674b146f071448373afbac49f8cdc5