Refactor ScalingLazyColumn classes so that scaling calculations are derived from changes to underlying LazyListState
Move the calculation of the scaling to apply to items in the ScalingLazyColumn so that it is derived from changes in the embedded LazyListState object. This reduces the number of scaling calculations being done.

The change also allows for ScalingLazyColumnItemInfo and ScalingLazyColumnLayoutInfo interfaces to be added to expose both the unadjusted (unscaled) offsets and sizes and the adjusted (scaled) offsets and sizes to be made available on the state.

A by-product of this refactoring is that 'top' content padding is now correctly handled for a ScalingLazyColumn - previously providing top padding would result in incorrect scaling of the items in the viewport.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 194464927
Bug: 194370448
Bug: 193792848

RelNote: "We have added ScalingLazyColumnItemInfo and ScalingLazyColumnLayoutInfo interfaces to ScalingLazyColumnState class to enable developers to know the actual positions and sizes of items in the ScalingLazyColumn after scaling has been applied. We have also fixed a bug with the way that scaling was calculated when 'top' content padding is applied the ScalingLazyColumn. Finally we "
Change-Id: I27c07d0048f076da4d30e3b602f86fa417f1f1f9