Ensure the correct profile key in the transcoded baseline profiles
Bug: 195324392

This fix results in baseline profiles needing transcoding for P,Q,R devices. This might result in a very small slowdown in the amount of time a profile installation takes for those devices, but will result in those devices receiving a big increase in performance as a result of the baseline profile being installed/used in the first place.

Change-Id: I84e89c43ac526876c1165cf489662437f26e84ae