Hide checkers from the public API.
This is currently intended for some warning-level checks. They can't be
extended by developers at the moment, and we only have the one
accessibility check, so let's remove this from the public API for now.
We can add it back in later when we have more checks.

Bug: 194806274
Relnote: "Removed layout checkers from public API. These will always be
enabled, although will only raise a warning if a check fails."
Test: Existing tests pass

Change-Id: Ie9f2944b028e8a0ad00f36a7748eb31c828499cf