Remove need for setup/post-process scripts for simpleperf, and make it default
Bug: 158303822
Test: ./gradlew benchmark:benchmark-common:cC benchmark:benchmark-macro:cC
Test: ./gradlew compose:material:material-benchmark:cC -P android.testInstrumentationRunnerArguments.androidx.benchmark.profiling.mode=StackSampling -P android.testInstrumentationRunnerArguments.class=androidx.compose.material.benchmark.CheckboxesInRowsBenchmark#draw
Relnote: Switched to Simpleperf as default sampling profiler on API 29+

Simpleperf is now the 'default' stack sampling profiler, as it doesn't
need the simpleperf setup scripts to be used on API 29+.

This CL also renames the profilers for clarity going forward.

Lowered default sampling rate to reduce liklihood of interference and
overwhelming the measure thread.

Unified logic for temporarily overriding a setprop with new
PropOverride class.

Change-Id: Ic4b346221bb91dea1e7078e369f106f9e48be1e3