Add support for injecting mouse gestures
Adds a MouseInjectionScope similar to the TouchInjectionScope, an
implementation of its elementary methods in MultiModalInjectionScope,
and support in InputDispatcher and AndroidInputDispatcher.

A few platform independent abstractions for mice have been introduced
because they didn't exist yet: MouseButton and ScrollWheel. These have
been made @ExperimentalTestApi since I expect to replace them with the
production variants when they become available. For this reason, and the
fact that the API and/or behavior might change in response to the final
implementation of mouse support in Compose, all mouse injection support
has been marked @ExperimentalTestApi for the time being.

Bug: 190493367
Test: Added MouseEventsTest.kt
Relnote: "Added experimental support for injecting mouse events. Use
`performMouseInput` to start sending mouse events, or send mouse events
through the `Mouse` property of `MultiModalInjectionScope` during a
multi modal input gesture with `performMultiModalInput`. See
`MouseInjectionScope` for documentation of the available API."

Change-Id: Iaa4a82131f35d2d6376d8edaa1fac4807b27045d