Update for UnlockedOrientationTest flaky
The setOrientationNatural could take 1 second
to take effect and the waitForIdle() sometimes
unable to ensure the orientation happened.
The open Activity might be closed when the
orientation happens, the timeout would happen
if the test is still waiting for some signal
on the closed Activity.

This CL will call the setOrientationNatural()
a little bit earlier to ensure the testing
Activity.

Test: ImageCaptureUnlockedOrientationTest,
ImageAnalysisUnlockedOrientationTest

Change-Id: I5fdfd244b0fccc067bc965ac095cfb0f1c0d756a