First Version of the Wear Material Scaffold component.
The Scaffold provides automatic handling of PositionIndicator (such as scroll position or volume setting) which is based on state and styling provided through the ScaffoldState. This allows the developer to tell the scaffold which type of indicator to show for the main application content. The PositionIndicator is located at the right center of the screen and is either curved or straight depending on the watch type. The Scaffold also overlays a Vignette to allow for blurring of the screen edges when there is scrollable main content.

The Scaffold also provides a slot for the developer to provide a TimeText (component will follow in another patch) that allows for displaying the time and application status in an area at the top of the watch. The TimeText component will be watch shape aware.

In order to be multi-platform capable an indirection via a new Resources.kt file is provided to enable the Vignette images to be loaded. Thanks to Louis (lpf@) and Jim (jsproch@) for advice on how to achieve this.

Thanks to ssancho@ for the original version of the PositionIndicator.kt code

Test: androidx.wear.compose.material.ScaffoldTest,androidx.wear.compose.material.PositionIndicatorTest
Bug: 196387222
RelNote: "We have added Scaffold a top level application composable that provides a structure for handling PositionIndicators (like Scroll or Volume) position, an area at the top of the screen for displaying the time and application status, and also supports a Vignette to blur the screen top and bottom edge for scrollable content. The main area of the Scaffold is where the application contents are placed."
Change-Id: I5e0bf5d3b7fe6ebc247a8b9ba255b00167fe35c4