Correct flaky coroutine test
Previously, the test relied on the job being actually cancelled, which
happens asynchronously from the test. It worked most of the time, but
may happend after the cancellation test happened.

Instead, we now open a latch when the job is cancelled, ensuring there
is no synchronization issues.

Fix: 197298087
Test: ./gradlew glance:glance-appwidget:test
Relnote: n/a
Change-Id: Id0f3cd522edd6669d544a27d33105b31308d3013