Workaround for projectDir conflicts
Points the automatically generated parent project's projectDir to a
non-existent file to workaround Gradle limitations around conflict
project roots. Normally, the automatically generated parent project
would default to its relative dir, but since it conflicts with the root
project in Playground builds, it reverts to
`frameworks/support/parent/parent`, which conflicts with a real project
in some playgrounds. This causes the AndroidXPlugin to get applied to
the parent project incorrectly, but also causes adverse affects in
studio's code completion and conflicts with the child project if it
exists.

Fixes: 197253160
Test: cd activity && ./gradlewtasks