Compress the output image with the specified JPEG quality level for capture modes when cropping is needed
Relnote: "When cropping is needed for ImageCapture, compressing the output cropped image with the specified JPEG quality level according to the set capture mode. If the capture mode is CAPTURE_MODE_MINIMIZE_LATENCY, the JPEG compression quality will be 95. If the capture mode is CAPTURE_MODE_MAXIMIZE_QUALITY, the JPEG compression quality will be 100."

Bug: 142856426
Test: ./gradlew camera:camera-core:build && ./gradlew camera:camera-core:connectedAndroidTest
Change-Id: Ieb37cfede4173d165bd481b86a04545887b0d487