Declare AndroidXPlaygroundRootPlugin in buildSrc
This PR fixes a bug in playground build where it fails to find
SDK location because it fails to detect the playground plugin.

When a plugin is not declared in the metadata, it does not get
and ID, hence the  check fails. I've added the metadata
and also changed the ID key that we use to search.

Bug: n/a
Test: gw :datastore:datastore-core:generateApi