Fix up PositionIndicatorTest.kt so that we are not mixing Dp and Px values to avoid rounding errors on certain density devices.
PositionIndicatorTest.kt was subject to rounding error due to itemHeight being defined in Px and converted to Dp and defaultItemSpacing being defined in Dp and converted to Px. On API 28 Pixel 2 device the expected and actual viewport height differed by 1 pixel. To avoid this the test has been rewritten with both item height and spacing defined in Px and converted to Dp.

As a result the item spacing has to be explicitly set to the dp value converted from Px.

Test: androidx.wear.compose.material.PositionIndicatorTest
Bug: 197817612
Change-Id: I12db917ec714d6a7488e7517da7438939e2d7eca