CameraController#clearPreviewSurface unbinds its own use cases only instead of calling unbindAll
When multiple CameraController and PreviewView are used in one Activity, the previous PreviewView might receive the onDetachedFromWindow callback after the new pair of CameraController and PreviewView are created. It will cause the previous CameraController to call unbindAll in clearPreviewSurface function. This will unbind the use cases bound by the new CameraController. CameraController should only unbind its own use cases. Otherwise, it might unexpectedly unbind use cases that belongs to another LifecycleCamera.

Relnote: "Fixed the issue that camera is closed unexpectedly when multiple CameraController and PreviewView are used in one Activity."

Bug: 197539358
Test: buildview='./gradlew camera:camera-view:build' && testview='./gradlew camera:camera-view:connectedAndroidTest'
Change-Id: Ibfd1842890b683ae3db639bf0cc4a2202110fe9f