Generate better swipes with velocity
This uses a new algorithm to derive a function for the path of the swipe
which results in much looser constraints. When a function can't be
created because the parameters are too strict, it will give suggestions
for adjustments in the parameters so that it becomes possible to
generate a swipe that ends with a given velocity.

Bug: 182477143
Test: ./gradlew compose:ui:ui-test:cC
Relnote: "Improved `TouchInjectionScope.swipeWithVelocity`. It now
accepts a wider range of input variables and will suggest changes to the
input if a swipe can't be created"

Change-Id: I40fbe57b6f28689bc2597b2e64ae0ef28158e593