Correct the text weight handling.
It turns out that, if the font family is not set, then the text font
weight is ignored. Here, I add the font family (sans-serif, which is the
default) ... with the full knowledge this is going to be a problem when
we'll want to allow setting the family.

Relnote: n/a
Test: ./gradlew glance:glance-appwidget:connectedTest
Bug: 198819119
Change-Id: Ic44c68e0a9317cfa1b233748e772112907b7c6e0