[GH] Update gradle caching
gradle-command-action action uses the arguments as the cache key and restores dependency cache only in the first run.
This means, it uses the "changed files" as the cache key, which is the same for all projects, making it useless as they'll override each-other.

This PR adds a first invocation to the action and passes the project name as an argument + the toml file.
This should allow each sub project to have its own cache.

There is a certain risk that this might hit the 5GB cache limit as projects won't share cache, causing them to evict each-other (instead of overriding each-other). We don't force exact so it should at least recover from prefix cache key.

Whether this change would help or not is questionable but if we don't get into a cache-eviction race, it might work well.

To be explicit, i disabled the cache for subsequent calls to gradle-command-action and also updated it back to @v1  as the previous issue we hit is fixed.
https://github.com/gradle/gradle-build-action/issues/70

Bug: n/a
Test: CI

This is an imported pull request from https://github.com/androidx/androidx/pull/242.

Resolves #242
Github-Pr-Head-Sha: a2189365895ac48b4aee689b587913f788084b70
GitOrigin-RevId: c0353a0b198142404e05eded654849cbe81d3d5d
Change-Id: I0d9663c84b4f5a29c21da8bd23f1e0811e67d905