[GH] [XProcessing] Fix getAllAnnotations returning empty list
## Proposed Changes
There is a bug in `XAnnotated.getAllAnnotations` where it returns empty if the annotation contains a value named `value` and is an array type. This is due to a bug in `unwrapRepeatedAnnotationsFromContainer` where if the array it tries to unwrap is empty it is incorrectly believed to be a repeated annotations array.

The fix is simple, in `unwrapRepeatedAnnotationsFromContainer` we need to check that the array is not empty in addition to the existing `all` check we do.

## Testing

Test: Added a test to XAnnotationTest that reproduces the issue and passes with the fix applied.

## Issues Fixed

Fixes: N/A

This is an imported pull request from https://github.com/androidx/androidx/pull/243.

Resolves #243
Github-Pr-Head-Sha: ef5d30b730832bf54f78d54cd35fa85846e12731
GitOrigin-RevId: 175e12054cce96c96a8b1b486643fb8da992cd49
Change-Id: I70ad6e71e00c5488dfbf51d058d48dabf21de58c