Introduce ImageVectorCache local
Added caching mechanism to painterResource
API to cache all instances of ImageVectors that
are parsed with a given resource ID and theme.
Updated Xml parsing logic of vector graphics to
keep track of which configuration flags that
the vector asset depends on in order to invalidate
the cache on configuration changes.

Fixes: 178663631
Test: Added test to VectorTests to verify caching logic
Change-Id: Idf4197d4900b842fab0e3e658b061a482e7c6f80