Upload test apk's as a separate artifact
This PR fixes a problem where we stopped distributing test apks hence
stopped running FTL tests (aosp/1751630).

We will now re-use what aosp does to reduce chances of this breaking
(should probably add some failsafe, this PR does not do it).

For this to take affect, we also need to update androidx-ci-action to
look for the new artifact name.

Bug: n/a
Test: CI