Add sample-based metric result format
Fixes: 199940612
Test: ./gradlew benchmark:benchmark-common:cC benchmark:benchmark-macro:cC
Relnote: "Add new type of sampled metric to UI and JSON output, focused on
    percentiles of multiple samples per iteration."

Creates a second type of metric result, for percentile-focused metrics
like frame time. These metrics will record multiple values per
iteration, and display distributions (P50,90,95,99) from the entire
collected range of values.

This will drastically simplify the new frame metrics measurements,
both in terms of numbers of values reported in UI, but also in
avoiding showing distribution stats... of distribution stats, as we
did before.

Previous:

  frameTime50thPercentileMs   min   6.0,   median   6.0,   max   7.0
  frameTime90thPercentileMs   min   7.0,   median   7.0,   max   9.0
  frameTime95thPercentileMs   min   7.0,   median   7.0,   max   9.0
  frameTime99thPercentileMs   min  14.0,   median  14.0,   max  14.0
            totalFrameCount   min 294.0,   median 294.0,   max 294.0

New:

                frameTimeMs   P50  7.1,   P90  9.4,   P95  9.6,   P99  12.0

Note that this is only used in the FrameTimingTraceMetrics, so isn't
on by default yet.

Similarly, in JSON, we output this data in a new way, as a separate
"sampledMetric" map, e.g.:

    "sampledMetrics": {
        "frameTimeMs": {
            "P50": 7.072010000000001,
            "P90": 9.4366065,
            "P95": 9.62164825,
            "P99": 12.00158825,
            "runs": [
                [
                    16.025587,
                    8.140102,
                    7.137429,
                    7.380798,
                    ...
                ],
                ...
            ]
        }
    }

Change-Id: I56247e421890537d27f1e3f284209aa590472377