Fix back button handling when embedding a NavHost
When a destination in an outer NavHost contains
an inner NavHost and that destination gets put
on the outer NavHost's back stack, then you pop
back to that destination, the back button handling
gets messed up if that inner NavHost had its own
back stack, forcing users to pop the back stack
the same number of times that the inner NavHost
was pushed onto the back stack.

By tying the enabled state of the back handling
to the presence of the NavHost in composition,
we ensure that only visible NavControllers will
ever intercept the back button.

Test: new NavHostTest test passes
BUG: 195668143
Change-Id: Iffda7c8e12bde25a4cc687ac822afa3d8a7d0c57