Add density and layoutDirection to LayoutInfo
This exposes the Density and LayoutDirection of a node in its
LayoutInfo, which can be consumed for example in tests where assertions
and actions are done on nodes for which this information might be
necessary.

Fix: 200020607
Fix: 200019803
Test: ./gradlew compose:ui:ui:testDebug && \
        ./gradlew compose:ui:ui:cC
Relnote: "Added `density: Density` and `layoutDirection:
LayoutDirection` to `LayoutInfo`. This allows consumers of `LayoutInfo`
to interpret the dimensions and position exposed in `LayoutInfo`
properly."

Change-Id: I002f191f17be7fa867cecf9c974edfb694523786