Ensure we remove observers when removing menu Providers
If you add a MenuProvider using a Lifecycle owner, then call
removeMenuProvider before the Lifeycycle reaches DESTROYED, the observer
is never removed so there is an unneeded additional call to
removeMenuProvider that also invalidates the menu again.

We need to track all of the added observers and ensure they are also
removed when the call to removeMenuProvider is made.

RelNote: "MenuHostHelper will now properly remove all LifecycleObservers
when a MenuProvider is removed prior to the Lifecycle being DESTROYED."
Test: added MenuHostHelperTest
Bug: 199788262

Change-Id: Ibe2e1f4a29c8079549814aaaa9b2484c22a41b2e