Add lint rule for right lifecycle in addMenuProvider
The new MenuHost addMenuProvider APIs take a LifecycleOwner and for
fragments, they should pass it the viewLifecycleOwner instead of
themselves.

This new lint rule identifies whether a fragment calls addMenuProvider
on a class that extends MenuHost and if the fragment attempts to pass
itself as the LifecycleOwner is notifies the user they should be using
the fragment's viewLifecycleOwner instead.

Also fixed an issue in the FragmentLiveDataObserveDetectorTest where it
was passing improperly when a fragment was being used as the
LifecycleOwner to observe data.

RelNote: "Lint will now warn you when attempting to pass a fragment as
the LifecycleOwner to addMenuProvider methods."
Test: added AddMenuProviderDetectorTest
Bug: 200326272

Change-Id: I1813147e1db42b557942955725b50acabe8e49f6