Fix argument parameters generated in wrong order
With the recent change to ensure Kotlin always takes default parameters
last, we also needed to make sure the functions that create the
generated class (i.e. fromBundle/SavedStateHandle) generate arguments in
the proper order as well.

RelNote: "SafeArgs now generates the arguments for `fromBundle()` and
`fromSavedStateHandle()` in the proper parameter order."
Test: ./gradlew --rerun-tasks navigation:navigation-safe-args-generator:test
Bug: 200059831

Change-Id: I824a871363ff0aaf5fa5ef608f103582e7abfc61