Update XBasicAnnotationProcessor to not retain deferred elements across rounds.
It is important not to retain elements across rounds as that will lead to equality checks failing and obscure bugs. This change updates basic annotation processor implementations to follow the same strategy as auto-common for storing and finding deferred elements. That is, for each deferred element find its closest enclosing type element and store its FQN. In the next round use the name to find the type element from the environment and search its members for annotations requested by the steps.

The approach is not perfect, it is possible for an already processed and non-deferred element to be re-processed by a step which could potentially lead to a duplicate source being generated. This can be improved in a follow-up, but for now using the same approach as auto-common should cause no issues since that is what we've been using all this time.

Test: XProcessingStepTest.kt
Change-Id: If4ce5331e783db5e30b8fd9d246d585b0c0eec7d