Rework the AppCard and TitleCard to rename and move the body composable.
AppCard and TitleCard: Rename body to content, rename bodyColor to contentColor, move content to the be the last property to allow it to be provided as a lambda.
Update the documentation to make it clear that if the expected content of Text or Image is replaced by other composables that it is the responsibility of the developer to specify how to lay them out.

Bug: 197895820
Test: androidx.wear.compose.material.CardTest
RelNote: "We have renamed AppCard and TitleCard 'body' properties to 'content' and moved them to the end of the property list to allow them to be provided as a trailing lambda. Also renamed 'bodyColor' to 'contentColor' for consistency with the new slot names."
Change-Id: I57e78a7565904d5c32f7ee39888ea48bcf3478e8