Desktop. Fix returning wrong paragraph parameters
Fixes https://github.com/JetBrains/compose-jb/issues/1210

Compose uses the same intrinsics instance for two paragraphs internally.

When the second paragraph is created and layouted it messes all parameters for the first paragraphs.

It is because intrinsics uses the same instance of SkiaParagraph

The simplest solution - is to create a new SkiaParagraph every time:
1. in intrinsics, to compute min/max intrinsics width
2. in DesktopParagraph 1
3. in DesktopParagraph 2

The more complex solution is to reuse SkiaParagraph created in intrinsics (it is 2.5x faster)

That is how it was before, but now we reuse it only for the first DesktopParagraph, the second will create its own internal SkiaParagraph

Change-Id: I8def766d55cd184d8db5bcc89033de99fa5be6d5
Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true