Support dropping page cache without root on API 31/S+
Bug: 200160030
Test: ./gradlew benchmark:benchmark-common:cC benchmark:benchmark-macro:cC
Test: TrivialStartupBenchmark # comparing new vs old way to drop caches on API 31
Relnote: "Support dropping page cache without root on API 31/S+"

Additionally, with this change, failure to drop page cache will throw
on a rooted build, as it may be necessary to root the adb session (run
`adb root`). I noticed this because, when attempting to verify the new
numbers, the old approach was failing on my device.

In the future, we should explore alternative approximations
(e.g. writing/reading large amounts of disk contents to flush the
cache) for older APIs.

Change-Id: Iecfdbd7b712ceaba109c7551c669845db1ff51b7