De-flake load_multipleQueryArguments in LimitOffsetPagingSourceTest
The way this test was setup never guaranteed that .load() would actually
dispatch, and had a chance to deadlock when it did due to usage of
queryDispatcher and a single-threaded executor for queryExecutor.

Unfortunately, there isn't really a good way to guarantee this since we
need wait a bit after the job has started, and don't want to change
LimitOffsetPagingSource just for a test, so we use a 50ms sleep instead.

Fixes: 202322488
Test: ./gradlew room:room-paging:cC
Change-Id: Ic3cee676e2ec5f1750a25a2610267eb627ba7224