Fix list pane sizing in AbstractListDetailFragment
When using AbstractListDetailFragment, it is
important that any layout_ attributes added to
the list pane that is inflated are passed
directly to the SlidingPaneLayout so that
they are respected (notably, layout_width for
setting a minimum width before going into two pane
mode and layout_weight for setting the relative
weight between the list and detail pane).

By removing the intermediary FrameLayout and
directly adding the list pane to the SlidingPaneLayout,
we fix not only the single pane case (where previously
the fixed layout_width would either cause the right
edge to not match the right edge of the SlidingPaneLayout
or would cause the right edge to be off the screen),
but also the two pane case.

Relnote: "Fixed an issue with `AbstractListDetailFragment`
that caused the `layout_width` and `layout_weight` attributes
on the list pane returned by `onCreateListPaneView()` to
be incorrectly handled or ignored."
Test: tested in the integration-tests app
BUG: 200946592

Change-Id: I8c23391f2102dd2131a11438706501e8b85d09f0