Launch CDC activity on API 29 for biometric + credential
Addresses an issue with some devices on API 29 where invoking
BiometricPrompt for authentication with setDeviceCredentialAllowed(true)
will result in BIOMETRIC_ERROR_USER_CANCELED when the user attempts to
actually use PIN, pattern, or password. This is done by launching the
confirm device credential Settings activity on affected devices, which
should still present a prompt to the user if biometrics are enrolled on
API 29.

Because this fix relies on non-API features and implementation details
that may differ across devices, this fix is gated behind an allowlist of
vendors for which this problem is known to occur and for which the fix
has been verified.

Test: Manually verified using integration test app on affected devices
Test: ./gradlew biometric:biometric:test
Test: ./gradlew biometric:biometric:connectedAndroidTest
Test: ./gradlew biometric:integration-tests:testapp:connectedAndroidTest

Bug: 191617511
Change-Id: Ib9b88edcaf5a5e772b04e175296712929ab3cc56