Use Application Context to get Sidecar.
Use application context to get Sidecar. Using a generic context can
lead to memory leaks. For example using an Activity that doesn't handle
configuration changes will lead to a memory leak when rotating a device.
The reason is that Sidecar captures the Context that is passed in.

Relnote: Fix memory leak.
Bug: 202989046
Bug: 202250276
Test: manual - Use an Activity that ignores config changes.
Test: Install leak canary.
Test: Subscribe to WindowLayoutInfo changes.
Test: Roteate device a few times.
Change-Id: I3fc799fa458528a5cb18dac373aa024f99fa5a6a