Fix enter/exit events
Bug: 204597567

Stylus support hover and doesn't have buttons to indicate
that a "press" was done when the ACTION_HOVER_EXIT is
received prior to the ACTION_DOWN. This means that we
need to be clever when receiving an ACTION_HOVER_EXIT
so that we don't accidentally send an Exit when the
stylus presses. We delay the ACTION_HOVER_EXIT until
another MotionEvent is received or the post() is received.

There was a bug in which child pointer input handlers were
replaced when they shouldn't be, thus doing extra work.

Test: new tests. manual testing
Change-Id: Ia53a4c79018caf9d7e267191de8b84419c357ecf