Dismiss already popped DialogFragments
When you navigate() to a DialogFragment
destination and immediately popBackStack()
(either by synchronously navigating to a
non-FloatingWindow fragment destination
or by manually calling popBackStack()), the
popBackStack() call happens before the
DialogFragment has even been attached, much
less reached onCreate().

This means that the DialogFragment doesn't
actually get dismissed from the call to
popBackStack(). By checking the current
back stack in onCreate() of the DialogFragment,
we can ensure that we catch these dialogs
which previously slipped through the system
and manually dismiss them, thus ensuring
that the FragmentManager and visible state
match the state of the DialogFragmentNavigator.

Relnote: "Fixed a race condition where
popping a `DialogFragment` destination off of the
back stack before the dialog was actually shown
would not actually dismiss the dialog, resulting in
a crash when the errant dialog was manually
dismissed by the user."
BUG: 157679488
Test: new test passes

Change-Id: I687e5c540686b7eedf96a2c8875e6e456840ca69