Camera Logger tweaks
For the logging methods without a Throwable argument call the
corresponding Log methods rather than ultimately call the variant with
a Throwable that is set to null. The latter log methods are more
expensive (more allocation, more code).

Re-order checks in Logger.truncateTag so the cheap and likely to fail
test is first.

Avoid truncating tags twice on older slower devices.

Fix: 203975581
Test: Treehugger
Change-Id: Id070c9935f3e613d038bdb48a7c0f06bc49555f8