Fix race condition in shared paging data
This CL fixes a race condition in CachedPageEventFlow and also a memory
leak as a side effect :).

Previously, if source dispatches 2 or more events between the time new
subscriber connects to the history and connects to the shared flow, it
could miss events.

An earlier version of this CL does have a repro test but it is removed
now as this CL completely removes the two flow collection setup that we
had before (hence the case cannot be reproduced).

Instead, CachedPageEventFlow now uses the onSubscribed extension
function on SharedFlow to be able atomically inject a new stream.
onSubscribed is guaranteed to run between the time subscription is
established and before any other event is dispatched, hence we don't
have the race anymore.

As part of this change, CachedPageEventFlow will no longer use shareIn,
which also caused b/203594733. That leak happened because shared flows
never complete so it kept the CachedPageEventFlow in memory even when we
didn't use it anymore. Instead, this new implementation simply launches
a collection job in the given scope and sends the latest event
via a controlled MutableSharedFlow. The injection of history happens
on the onSubscribe of that MutableSharedFlow, which gives us an atomic
ability to inject the history before sending any more events.

This MutableSharedFlow has replay=1 and buffer=UNLIMITED. Buffer is only
ever used when there is a slow collector, which was still necessary in
the previous implementation as well.

To cancel the downstream, we still emit a complete event to this Flow.
Unlike the previous implementation, this is not a `active.flatMap`
operation anymore, instead we use `tryEmit` and rely on the fact that it
will never fail in practice (due to UNLIMITED buffer).

Bug: 204228426
Bug: 203594733
Test: CachedPageEventFlowTest (the memory leak test is in a followup)

Change-Id: Ib682ea334e061bc5fa850663cc1ebea0c7a7ebbf