Fix action parameters equality in PendingIntent
Android compares PendingIntents to decide if it needs to update them. To ensure we create different PI, we convert the provided Intent into a URI and assigned to the data value.

The problem was that when nesting the parameters inside a bundle key the generated URI was always the same, not updating the PI.

Create a random UUID to ensure everytime we create a different PI.

Test: ./gradlew :glance:glance-appwidget:test
Change-Id: Id586729ed1f445c63310ef9f6ae3730d37d86511